Changed ExtensionField to  QuadExtension
This PR is a minor change in the naming of the specific extension we
use. Builds on the PR by @klewi and @irakliyk for #79.

Minor update,found another reference.

Removed changes in fib2 and ran cargo commands

Fixed fmt error

Fixed reverting to original version of the file which showed an error in cli