Removes the `PositiveInteger = u128` constraints
This constraint allowed some implementations to use associated type constraints to require conversion of some configuration parameters (e.g. size of domain) into the scalar types that we use (== u128).
Yet :
- this did limit the usability of our implementation (in particular, a lot of code became unusable with u64 scalar size),
- this makes generic parameters quite unwieldy.

In general, associated type constraints should not be used unless there's a spectacular, local reason for it.

The current PR removes the constraint and instead obtains this convertibility from the `From<u32>` constraint for the exponent domain.
We should consider :
- rebasing #79 on the present PR,
- specifying a few of the existing parameters as u32 rather than usize.